Modify xmltohb.pl to support 4 byte attribute IDs
RTC: 45610
Reviewed-on: http://gfw160.austin.ibm.com:8080/gerrit/1565
Tested-by: Jenkins Server

+++ src/usr/targeting/common/target.C
@@ -1,26 +1,25 @@
-/*  IBM_PROLOG_BEGIN_TAG
- *  This is an automatically generated prolog.
- *
- *  $Source: src/usr/targeting/common/target.C $
- *
- *  IBM CONFIDENTIAL
- *
- *  COPYRIGHT International Business Machines Corp. 2011-2012
- *
- *  p1
- *
- *  Object Code Only (OCO) source materials
- *  Licensed Internal Code Source Materials
- *  IBM HostBoot Licensed Internal Code
- *
- *  The source code for this program is not published or other-
- *  wise divested of its trade secrets, irrespective of what has
- *  been deposited with the U.S. Copyright Office.
- *
- *  Origin: 30
- *
- *  IBM_PROLOG_END_TAG
- */
+/* IBM_PROLOG_BEGIN_TAG                                                   */
+/* This is an automatically generated prolog.                             */
+/*                                                                        */
+/* $Source: src/usr/targeting/common/target.C $                           */
+/*                                                                        */
+/* IBM CONFIDENTIAL                                                       */
+/*                                                                        */
+/* COPYRIGHT International Business Machines Corp. 2011,2012              */
+/*                                                                        */
+/* p1                                                                     */
+/*                                                                        */
+/* Object Code Only (OCO) source materials                                */
+/* Licensed Internal Code Source Materials                                */
+/* IBM HostBoot Licensed Internal Code                                    */
+/*                                                                        */
+/* The source code for this program is not published or otherwise         */
+/* divested of its trade secrets, irrespective of what has been           */
+/* deposited with the U.S. Copyright Office.                              */
+/*                                                                        */
+/* Origin: 30                                                             */
+/*                                                                        */
+/* IBM_PROLOG_END_TAG                                                     */
 /**
  *  @file targeting/common/target.C
  *
@@ -234,7 +233,7 @@ uint8_t * Target::targetFFDC( uint32_t & o_size ) const
                             sizeof(attrClass) + sizeof(attrType) + 
                             sizeof(attrModel);
 
-    uint8_t attrEnum = ATTR_NA;
+    uint32_t attrEnum = ATTR_NA;
     
     uint8_t pathPhysSize = 0;
     AttributeTraits<ATTR_PHYS_PATH>::Type pathPhys;
